requirements.txt
certifi==2020.6.20
cffi==1.14.0
cryptography==2.9.2
decorator==4.4.2
emoji==0.5.4
pkg-resources==0.0.0
pycparser==2.20
PySocks==1.7.1
python-telegram-bot==12.8
six==1.15.0
tornado==6.0.4
